Add structured refs readback recommendations
